fix(asr): force-link host libstdc++ via cargo:rustc-link-arg for zigbuild
Iteration #2 (cd5b7eb9) confirmed zig's hermetic lld IGNORES host
cargo:rustc-link-search paths for the C++ stdlib: -L/usr/lib/gcc/.../13 and
-lstdc++ were both on the link line yet the link still failed with the same
undefined std::* symbols (std::string::_Rep::_S_empty_rep_storage). The
link-search approach (fae-engine/build.rs) is dead — removed it.

Iteration #3: resolve the host gcc's libstdc++ to an ABSOLUTE path and emit it as
cargo:rustc-link-arg=<path> in fae-daemon/build.rs. lld links the exact file
verbatim, bypassing sysroot search entirely. cargo:rustc-link-arg does NOT
propagate from a library crate, so this lives in fae-daemon (the binary being
linked). Prefer the self-contained static .a (no runtime .so dep, no
linker-script parsing); fall back to .so. Harmless on native gcc; no-op on macOS.

GNU libstdc++ (NOT libc++): the undefined symbols are GNU-libstdc++ ABI which
LLVM libc++ does not export.
